Fix `evm-tracing-events` integrations
